BJS-15677
