--- src/main/java/faang/school/achievement/config/redis/RedisConfig.java
@@ -1,39 +1,0 @@
-package faang.school.achievement.config.redis;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-public class RedisConfig {
-
-    @Value("${spring.data.redis.channel.comment_achievement}")
-    private String commentChannel;
-    @Value("${spring.data.redis.channel.achievement}")
-    private String achievement;
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory connectionFactory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(connectionFactory);
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(new Jackson2JsonRedisSerializer<>(Object.class));
-
-        return template;
-    }
-
-    @Bean
-    public ChannelTopic achievementTopic() {
-        return new ChannelTopic(achievement);
-    }
-
-    @Bean
-    public ChannelTopic commentChannel() {
-        return new ChannelTopic(commentChannel);
-    }
-}
