+++ src/main/java/faang/school/achievement/model/Achievement.java
@@ -1,5 +1,6 @@
 package faang.school.achievement.model;
 
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -37,6 +38,7 @@ public class Achievement {
     private List<UserAchievement> userAchievements;
 
     @OneToMany(fetch = FetchType.EAGER, mappedBy = "achievement")
+    @JsonManagedReference
     private List<AchievementProgress> progresses;
 
     @Column(name = "points", nullable = false)
